wvstreams_unittest: ported all recent changes from HEAD into this branch, then set wvstreams_unittest_lastmerge to point at current HEAD.  (This is what I should have done all along, but I was changing wvstreams_unittest_branchpoint instead.  Be careful!)  Also, forgot to check that in before I started munging other stuff to make xplc work better with UniConf.
git-svn-id: file:///home/apenwarr/alumnit-svn/public/branches/wvstreams_unittest@3253 6619c942-59fa-0310-bf43-a07aaa81d9ac

+++ include/wvbdbhash.h
@@ -165,7 +165,7 @@ public:
     bool exists(const K &key)
         { return WvBdbHashBase::exists(datumize<K>(key)); }
 
-    int count()
+    size_t count()
     {
 	int res = 0;
 	Iter i(*this);
